backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
python-multipart>=0.0.9,<1.0
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.29,<1.0
pydantic-settings>=2.2,<3.0
redis>=5.0,<6.0
orjson>=3.10,<4.0
langchain>=0.3,<0.4
langgraph>=0.2,<0.3
langchain-openai>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-text-splitters>=0.3,<0.4
faiss-cpu>=1.8,<2.0
pypdf>=5.0,<6.0
python-docx>=1.1,<2.0
tiktoken>=0.8,<1.0
httpx>=0.27,<1.0
